+++ src/mc_tasks/CoMTask.cpp
@@ -10,8 +10,8 @@ CoMTask::CoMTask(const mc_rbdyn::Robots & robots, unsigned int robotIndex)
 
   cur_com = rbd::computeCoM(robot.mb(), robot.mbc());
 
-  comTask.reset(new tasks::qp::CoMTask(robots.mbs(), robotIndex, cur_com));
-  comTaskSp.reset(new tasks::qp::SetPointTask(robots.mbs(), robotIndex, comTask.get(), 5, 100));
+  comTask.reset(new tasks::qp::CoMTask(robots.mbs(), static_cast<int>(robotIndex), cur_com));
+  comTaskSp.reset(new tasks::qp::SetPointTask(robots.mbs(), static_cast<int>(robotIndex), comTask.get(), 5, 100));
 }
 
 void CoMTask::resetTask(const mc_rbdyn::Robots & robots, unsigned int robotIndex)
